Remove non data-structure code
The primary purpose of this project was testing code in userspace before
pushing it to the x15 microkernel. The xprintf and phys modules were
included quite some time ago now, along with the mem module which is also
present in the GNU Mach microkernel.

The secondary goal of this project was to create good implementations of
commonly used data structures, mainly the doubly linked list and the
red-black tree, under a BSD license so that they're reusable in proprietary
projects I have to work on. Such generic code remains in this repository.